Remove palette rotation code, load sprite palettes
The previous commit introduced loading a single sprite and also
introduced a table of palettes for sprites, but it did not load the
sprite palettes into the PPU, so the default gray-ish color for the
sprite was getting lost in the gray road behind it.

This commit introduces two fixes:
* Adds support for loading the sprite palettes, so the sprites have
colors other than gray-ish.
* Moves the background and sprite palettes back into ROM with the rest
of the code.  The excessive use of the limited ZEROPAGE RAM was starting
to overflow the allowed area, and the only reason for keeping the
palettes in ZEROPAGE was for making some easily-documented palette
rotation code.