Correct handling of sprite_x/y variables
This commit cleans up the handling of the sprite_x and sprite_y
variables.  The original code was adapted from a simple sprite loading
demo that didn't care about the Y coordinate very much and, as a result,
it had some not-completely-obvious code in it that actually just forced
the Y coordinate to a fixed constant every time the sprite was loaded.
This was fixed and load_sprite was corrected to ensure sprite_x and
sprite_y still refer to the correct origin coordinate of a 2x2 sprite on
return.

Additionally, sprite_x and sprite_y were being given constant values in
initialization, but they're in the ZEROPAGE, which is a BSS segment,
meaning they must instead default to 0.  So, to fix this, their defaults
are now set early in the code.