Fix loading the mushroom sprite
There were various issues with loading the mushroom sprite in the prior
commit, most notably being that the X register was getting reset to 16
repeatedly, which got in the way of using it as an array index for the
sprite page.  Additionally, many of the sprites were being loaded to the
same coordinates, so they'd have appeared on top of one another.

This commit cleans up those issues and moves load_mushroom to its own
proc, which is not strictly needed but does help a bit from a code
organization point of view.